# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TASK MANAGER API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/add/student": {
            "post": {
                "tags": [
                    "User API"
                ],
                "summary": "Register user",
                "operationId": "6cfeca03a31a2cf4850e1007957c7ab6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fullname": {
                                        "type": "string",
                                        "example": "Sarvarbek"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": "upload/avatar/file/kFSwDf0msLoCMWpyPAL4.jpg"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": "13"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Andijon"
                                    },
                                    "group_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "course_id": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration user and to login.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "username": "Abdullajon",
                                                "password": "$2y$10$sZZUV7ykVyAXT7/eDzGyReThtdkaLUK9jknrDEDNaeK71uKeQkfiO",
                                                "token": "ms7ST2opzINdvWCxLanr8zg21kOKFS",
                                                "updated_at": "2023-08-28T11:58:12.000000Z",
                                                "created_at": "2023-08-28T11:58:12.000000Z",
                                                "id": 1
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Login!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/update/student/{id}": {
            "post": {
                "tags": [
                    "User API"
                ],
                "summary": "Register user",
                "operationId": "940c3fd21488f30614740d9e4539632f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Paste task id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fullname": {
                                        "type": "string",
                                        "example": "Sarvarbek 2"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": "upload/avatar/file/kFSwDf0msLoCMWpyPAL4.jpg"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": "13"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Andijon"
                                    },
                                    "group_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "course_id": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration user and to login.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "username": "Abdullajon",
                                                "password": "$2y$10$sZZUV7ykVyAXT7/eDzGyReThtdkaLUK9jknrDEDNaeK71uKeQkfiO",
                                                "token": "ms7ST2opzINdvWCxLanr8zg21kOKFS",
                                                "updated_at": "2023-08-28T11:58:12.000000Z",
                                                "created_at": "2023-08-28T11:58:12.000000Z",
                                                "id": 1
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Login!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/delete/student/{id}": {
            "delete": {
                "tags": [
                    "User API"
                ],
                "summary": "Register user",
                "operationId": "a37e900e7a1404228cc7f214fd334582",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Paste task id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fullname": {
                                        "type": "string",
                                        "example": "Sarvarbek"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "example": "upload/avatar/file/kFSwDf0msLoCMWpyPAL4.jpg"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": "13"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Andijon"
                                    },
                                    "group_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "course_id": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration user and to login.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "username": "Abdullajon",
                                                "password": "$2y$10$sZZUV7ykVyAXT7/eDzGyReThtdkaLUK9jknrDEDNaeK71uKeQkfiO",
                                                "token": "ms7ST2opzINdvWCxLanr8zg21kOKFS",
                                                "updated_at": "2023-08-28T11:58:12.000000Z",
                                                "created_at": "2023-08-28T11:58:12.000000Z",
                                                "id": 1
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Login!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}